ecommerce/ecommerce_requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Limiter==2.4.0
PyJWT==2.3.0
bcrypt==3.2.0
pyOpenSSL==20.0.1  # For SSL support
SQLAlchemy==1.4.23
python-dotenv==0.19.0
